libc2dcolorconvert: add RTLD_NODELETE flag when dlopen
Add RTLD_NODELETE flag to avoid unload the library during dlclose(),
Consequently, the library's static variables are not reinitialized
if the library is reloaded with dlopen() at a later time.
Consider a scenario when scroll up/down the playlist of mediacenter,
there will be lots of dlopen/dlclose, and some static variables in
those libs will be reinitialized, this is not our expected behaivor.

Change-Id: I99d05bf6685f4fddd12d650c1b3fabf98aaed0a3
Signed-off-by: Hui Li <hul@codeaurora.org>